// src/api/generated/models/addressResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { AddressResourceComplement } from './addressResourceComplement';
import type { AddressResourceAddressable } from './addressResourceAddressable';
import type { AddressResourceCreatedAt } from './addressResourceCreatedAt';
import type { AddressResourceUpdatedAt } from './addressResourceUpdatedAt';

export interface AddressResource {
  id: string;
  cep: string;
  address_type: string;
  street: string;
  street_complete_name: string;
  district: string;
  city: string;
  state: string;
  latitude: string;
  longitude: string;
  number: string;
  complement: AddressResourceComplement;
  addressable_id: string;
  addressable?: AddressResourceAddressable;
  created_at: AddressResourceCreatedAt;
  updated_at: AddressResourceUpdatedAt;
}